requirements.txt
# API 框架
fastapi==0.110.0
uvicorn==0.27.1

# HTTP 客戶端
aiohttp==3.9.3
requests==2.31.0

# 環境變數管理
python-dotenv==1.0.1

# 資料驗證
pydantic==2.6.3

# 型別提示支援
typing-extensions==4.10.0
